Feature/join group

+++ src/api/group/group.service.ts
@@ -31,6 +31,22 @@ class GroupService {
         }
         return mapTo(await this.repository.create(newGroup))
     }
+    async editGroup(groupId: string, callerId: string, groupUpdated?: any): Promise<IGroup> {
+        const group: IGroupDTO = await this.repository.getGroupById(groupId)
+        if (!group) {
+            console.log('Group not found')
+            throw GROUP_ERROR_CODE.GROUP_NOT_FOUND
+        }
+        if (
+            group.owner !== callerId?.toString() &&
+            !group.coOwners.includes(callerId?.toString())
+        ) {
+            console.log('Caller is not owner or co-owner')
+            throw GROUP_ERROR_CODE.NOT_HAVING_PERMISSION
+        }
+        const updatedGroup = await this.repository.updateById(groupId, groupUpdated)
+        return mapTo(updatedGroup)
+    }
     async getGroupHasMember(memberId: string): Promise<IGroup[]> {
         const pipeline = [
             {
@@ -143,10 +159,18 @@ class GroupService {
         return mapTo(await this.repository.updateById(groupId, group))
     }
 
-    async deleteGroup(groupId: string): Promise<IGroup> {
+    async deleteGroup(groupId: string, callerId: string): Promise<IGroup> {
         const group: IGroupDTO = await this.repository.getGroupById(groupId)
         if (!group) {
+            console.log('Group not found')
             throw GROUP_ERROR_CODE.GROUP_NOT_FOUND
+        }
+        if (
+            group.owner !== callerId?.toString() &&
+            !group.coOwners.includes(callerId?.toString())
+        ) {
+            console.log('Caller is not owner or co-owner')
+            throw GROUP_ERROR_CODE.NOT_HAVING_PERMISSION
         }
         return mapTo(await this.repository.deleteById(groupId))
     }
